+++ packages/time-utils/src/get-default-absolute-time/get-default-absolute-time.test.ts
@@ -1,40 +1,6 @@
 /* eslint-disable @typescript-eslint/no-this-alias */
 import { advanceTo, clear as clearDate } from 'jest-date-mock';
 
-const utcPlugin = jest.requireActual('dayjs/plugin/utc');
-
-/**
- * This is naive implementation of dayjs.utcOffset method.
- * It was created to make test deterministic and
- * prevent timezone DST values from affecting test results.
- */
-jest.mock(
-  'dayjs/plugin/utc',
-  () => (
-    option: string,
-    Dayjs: Record<string, any>,
-    dayjs: Record<string, any>
-  ) => {
-    const proto = Dayjs.prototype;
-    const plugin = utcPlugin(option, Dayjs, dayjs);
-    const hourOffset = 60;
-
-    proto.utcOffset = function (input: number, keepLocalTime?: boolean) {
-      const instace = this;
-      if (!input) return hourOffset;
-      if (input && keepLocalTime) {
-        instace.$offset = hourOffset;
-        instace.$u = input === 0;
-        return instace;
-      }
-
-      return instace;
-    };
-
-    return plugin;
-  }
-);
-
 import getDefaultAbsoluteTime from './get-default-absolute-time';
 
 beforeEach(() => {
@@ -46,11 +12,11 @@ afterAll(() => {
   clearDate();
 });
 
-test('get default timeframe for "Europe/Warsaw" timezone', () => {
-  expect(getDefaultAbsoluteTime('Europe/Warsaw')).toMatchInlineSnapshot(`
+test('get default timeframe', () => {
+  expect(getDefaultAbsoluteTime()).toMatchInlineSnapshot(`
     Object {
-      "end": "2021-03-31T00:00:00+01:00",
-      "start": "2021-03-30T00:00:00+01:00",
+      "end": "2021-03-31T00:00:00",
+      "start": "2021-03-30T00:00:00",
     }
   `);
 });
