+++ packages/ui-core/src/components/date-picker/date-picker.styles.ts
@@ -1,1116 +1,287 @@
-import styled, { createGlobalStyle, css } from 'styled-components';
+import styled, { createGlobalStyle } from 'styled-components';
+import { motion } from 'framer-motion';
 import { transparentize } from 'polished';
 import { colors } from '@keen.io/colors';
 
-export const inputMixin = () => css`
-  padding: 10px 14px;
-  height: 37px;
-  width: 100%;
-  box-sizing: border-box;
-  border: 1px solid ${transparentize(0.5, colors.black[500])};
-  border-radius: 4px;
-  outline: none;
-
-  font-size: 14px;
-  line-height: 17px;
-  font-family: 'Lato Regular', sans-serif;
-  color: ${colors.blue[500]};
-
-  &:focus {
-    box-shadow: 0 0 3px 1px rgba(119, 163, 187, 0.5);
-  }
-
-  &:disabled {
-    border: 1px solid ${transparentize(0.5, colors.blue[500])};
-    background: ${colors.white[500]};
-  }
-
-  &:disabled::placeholder {
-    color: ${transparentize(0.6, colors.black[400])};
-  }
-`;
-
 export const GlobalStyle = createGlobalStyle`
-body {
-  .rc-time-picker {
-    display: inline-block;
-    position: relative;
-    box-sizing: border-box;
-  }
-  .rc-time-picker * {
-    box-sizing: border-box;
-  }
+.react-calendar {
+  width: 320px;
+  max-width: 100%;
+  background: white;
+  border: none;
+  font-family: 'Lato', sans-serif;
+  line-height: 1.125em;
+}
+.react-calendar--doubleView {
+  width: 700px;
+}
+.react-calendar--doubleView .react-calendar__viewContainer {
+  display: flex;
+  margin: -0.5em;
+}
+.react-calendar--doubleView .react-calendar__viewContainer > * {
+  width: 50%;
+  margin: 0.5em;
+}
+.react-calendar,
+.react-calendar *,
+.react-calendar *:before,
+.react-calendar *:after {
+  box-sizing: border-box;
+}
+.react-calendar button {
+  margin: 0;
+  border: 0;
+  outline: none;
+  font-family: 'Lato', sans-serif;
+}
+.react-calendar button:enabled:hover {
+  cursor: pointer;
+}
+.react-calendar__navigation {
+  display: flex;
+  height: auto;
+  padding: 10px 20px;
+  border-bottom: 1px solid ${colors.white[300]};
+  column-gap: 20px;
+}
+.react-calendar__navigation button {
+  background: none;
+  padding: 0;
+  min-width: 20px;
+}
+.react-calendar__navigation button:enabled:hover,
+.react-calendar__navigation button:enabled:focus-visible {
+  background-color: #e6e6e6;
+}
+.react-calendar__navigation button[disabled] {
+  background-color: #f0f0f0;
+}
+.react-calendar__month-view__weekdays {
+  text-align: center;
+  color: ${transparentize(0.3, colors.black[500])};
+  font-size: 0.75rem;
+  font-weight: 400;
+  text-transform: none;
 
-  .rc-time-picker-input {
-    width: 100%;
-    position: relative;
-    display: inline-block;
-    padding: 4px 7px;
-    height: 28px;
-    cursor: text;
-    font-size: 12px;
-    line-height: 1.5;
-    color: #666;
-    background-color: #fff;
-    background-image: none;
-    border: 1px solid #d9d9d9;
-    border-radius: 4px;
-    transition: border 0.2s cubic-bezier(0.645, 0.045, 0.355, 1), background 0.2s cubic-bezier(0.645, 0.045, 0.355, 1), box-shadow 0.2s cubic-bezier(0.645, 0.045, 0.355, 1);
-  }
-  .rc-time-picker-input[disabled] {
-    color: #ccc;
-    background: #f7f7f7;
-    cursor: not-allowed;
-  }
-  .rc-time-picker-panel {
-    z-index: 1070;
-    width: 170px;
-    position: absolute;
-    box-sizing: border-box;
-  }
-  .rc-time-picker-panel * {
-    box-sizing: border-box;
-  }
-  .rc-time-picker-panel-inner {
-    display: inline-block;
-    position: relative;
-    outline: none;
-    list-style: none;
-    font-size: 12px;
-    text-align: left;
-    background-color: #fff;
-    border-radius: 4px;
-    box-shadow: 0 1px 5px #ccc;
-    background-clip: padding-box;
-    border: 1px solid #ccc;
-    line-height: 1.5;
-  }
-  .rc-time-picker-panel-narrow {
-    max-width: 113px;
-  }
-  .rc-time-picker-panel-input {
-    margin: 0;
-    padding: 0;
-    width: 100%;
-    cursor: auto;
-    line-height: 1.5;
-    outline: 0;
-    border: 1px solid transparent;
-  }
-  .rc-time-picker-panel-input-wrap {
-    box-sizing: border-box;
-    position: relative;
-    padding: 6px;
-    border-bottom: 1px solid #e9e9e9;
-  }
-  .rc-time-picker-panel-input-invalid {
-    border-color: red;
-  }
-  .rc-time-picker-panel-select {
-    float: left;
-    font-size: 12px;
-    border: 1px solid #e9e9e9;
-    border-width: 0 1px;
-    margin-left: -1px;
-    box-sizing: border-box;
-    width: 56px;
-    max-height: 144px;
-    overflow-y: auto;
-    position: relative;
-  }
-  .rc-time-picker-panel-select-active {
-    overflow-y: auto;
-  }
-  .rc-time-picker-panel-select:first-child {
-    border-left: 0;
-    margin-left: 0;
-  }
-  .rc-time-picker-panel-select:last-child {
-    border-right: 0;
-  }
-  .rc-time-picker-panel-select ul {
-    list-style: none;
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    width: 100%;
-  }
-  .rc-time-picker-panel-select li {
-    list-style: none;
-    margin: 0;
-    padding: 0 0 0 16px;
-    width: 100%;
-    height: 24px;
-    line-height: 24px;
-    text-align: left;
-    cursor: pointer;
-    user-select: none;
-  }
-  .rc-time-picker-panel-select li:hover {
-    background: #edfaff;
-  }
-  li.rc-time-picker-panel-select-option-selected {
-    background: #f7f7f7;
-    font-weight: bold;
-  }
-  li.rc-time-picker-panel-select-option-disabled {
-    color: #ccc;
-  }
-  li.rc-time-picker-panel-select-option-disabled:hover {
-    background: transparent;
-    cursor: not-allowed;
-  }
+  abbr {
+    text-decoration: none;
 
-  .keen-time-picker.rc-time-picker-panel {
-    font-family: 'Lato Regular', sans-serif;
-
-    .rc-time-picker-panel-input-wrap {
-      border: none;
-      padding: 0;
+    &[aria-label="Saturday"], &[aria-label="Sunday"] {
+      color: ${colors.blue[500]};
     }
+  }
+}
+.react-calendar__month-view__weekdays__weekday {
+  padding: 10px 5px;
+}
+.react-calendar__month-view__weekNumbers .react-calendar__tile {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 0.75em;
+  font-weight: bold;
+  padding: calc(0.75em / 0.75) calc(0.5em / 0.75);
+}
+.react-calendar__month-view__days__day.react-calendar__month-view__days__day--weekend {
+  color: ${colors.blue[500]};
+}
+.react-calendar__month-view__days__day--neighboringMonth {
+  opacity: 0.3;
+}
+.react-calendar__year-view .react-calendar__tile,
+.react-calendar__decade-view .react-calendar__tile,
+.react-calendar__century-view .react-calendar__tile {
+  padding: 10px 5px;
+  border-radius: 5px;
+  font-size: 14px;
+  border: 2px solid ${colors.white[500]};
+}
+.react-calendar__century-view .react-calendar__tile {
+  font-size: 13px;
+}
+.react-calendar__tile {
+  max-width: 100%;
+  text-align: center;
+  padding: 0.75em 0.5em;
+  background: none;
+}
+.react-calendar__tile:disabled {
+  background-color: #f0f0f0;
+}
+.react-calendar__tile:enabled:hover,
+.react-calendar__tile:enabled:focus-visible {
+  background-color: #e6e6e6;
+}
+.react-calendar__tile--now {
+  background: #ffff76;
+}
+.react-calendar__tile--now:enabled:hover,
+.react-calendar__tile--now:enabled:focus-visible {
+  background: #ffffa9;
+}
+.react-calendar__tile--hasActive {
+  background: #76baff;
+}
+.react-calendar__tile--hasActive:enabled:hover,
+.react-calendar__tile--hasActive:enabled:focus-visible {
+  background: #a9d4ff;
+}
+.react-calendar--selectRange .react-calendar__tile--hover {
+  background-color: #e6e6e6;
+}
 
-    .rc-time-picker-panel-input {
-      display: none;
+.react-calendar__month-view__days__day abbr {
+  font-size: 14px;
+  border-radius: 50%;
+  width: 37px;
+  height: 37px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.react-calendar__tile.react-calendar__month-view__days__day {
+  padding: 0;
+}
+
+.react-calendar__month-view {
+  padding: 10px 20px 20px 20px;
+}
+
+.react-calendar__tile--now,
+.react-calendar__tile:enabled:hover, .react-calendar__tile:enabled:focus-visible,
+.react-calendar__navigation__label:enabled:hover, .react-calendar__navigation__label:enabled:focus-visible,
+.react-calendar__navigation button:enabled:hover, .react-calendar__navigation button:enabled:focus-visible,
+.react-calendar__navigation button[disabled] {
+  background: ${transparentize(0.9, colors.blue[500])};
+}
+
+button.react-calendar__tile--active abbr,
+button.react-calendar__tile--active:enabled:hover abbr,
+button.react-calendar__tile--active:enabled:focus-visible abbr
+button.react-calendar__tile--active.react-calendar__month-view__days__day--weekend abbr,
+button.react-calendar__tile--active.react-calendar__month-view__days__day--weekend:enabled:hover abbr,
+button.react-calendar__tile--active.react-calendar__month-view__days__day--weekend:enabled:focus-visible abbr,
+.react-calendar__tile.react-calendar__tile--active.react-calendar__month-view__days__day:hover abbr,
+.react-calendar__tile.react-calendar__tile--active.react-calendar__month-view__days__day:focus-visible abbr,
+.react-calendar__tile--hasActive,
+.react-calendar__tile--hasActive:enabled:hover,
+.react-calendar__tile--hasActive:enabled:focus-visible,
+button.react-calendar__tile--hasActive.react-calendar__year-view__months__month,
+button.react-calendar__tile--hasActive.react-calendar__year-view__months__month:enabled:hover,
+button.react-calendar__tile--hasActive.react-calendar__year-view__months__month:enabled:focus-visible,
+button.react-calendar__tile--hasActive.react-calendar__decade-view__years__year,
+button.react-calendar__tile--hasActive.react-calendar__decade-view__years__year:enabled:hover,
+button.react-calendar__tile--hasActive.react-calendar__decade-view__years__year:enabled:focus-visible {
+  background: ${colors.blue[500]};
+  color: ${colors.white[500]};
+}
+
+button.react-calendar__tile--hasActive.react-calendar__year-view__months__month,
+button.react-calendar__tile--hasActive.react-calendar__year-view__months__month:enabled:hover,
+button.react-calendar__tile--hasActive.react-calendar__year-view__months__month:enabled:focus-visible,
+button.react-calendar__tile--hasActive.react-calendar__decade-view__years__year,
+button.react-calendar__tile--hasActive.react-calendar__decade-view__years__year:enabled:hover,
+button.react-calendar__tile--hasActive.react-calendar__decade-view__years__year:enabled:focus-visible {
+  border: 2px solid ${colors.white[500]};
+}
+
+.react-calendar__navigation__label__labelText {
+  font-size: 1rem;
+  color: ${colors.blue[500]};
+}
+
+button.react-calendar__navigation__arrow:enabled:hover,
+button.react-calendar__navigation__arrow:enabled:focus-visible {
+  background: none;
+}
+
+button.react-calendar__navigation__arrow:enabled:focus-visible {
+  outline: 1px solid ${transparentize(0.5, colors.blue[500])};
+}
+
+button.react-calendar__navigation__label {
+  padding: 10px;
+  border-radius: 4px;
+  transition: background-color 250ms ease-in-out;
+}
+
+.react-calendar__year-view__months,
+.react-calendar__decade-view__years,
+.react-calendar__century-view__decades {
+  padding: 20px;
+}
+
+.react-calendar__viewContainer {
+  min-height: 205px;
+}
+
+.react-calendar__tile.react-calendar__month-view__days__day:hover,
+.react-calendar__tile.react-calendar__month-view__days__day:focus-visible {
+    background: none;
+
+    abbr {
+      background: ${transparentize(0.9, colors.blue[500])};
     }
+  }
 
-    .rc-time-picker-panel-inner {
-      margin-top: 42px;
+.react-calendar__tile--now.react-calendar__month-view__days__day {
+  background: none;
+  abbr {
+    border: 1px solid ${transparentize(0.5, colors.blue[500])};
+  }
+
+  &:hover,
+  &:focus-visible {
+    abbr {
+      background: ${transparentize(0.9, colors.blue[500])};
     }
+  }
+}
 
-    .rc-time-picker-panel-select li {
-      outline: none;
-    }
+.react-calendar__tile--now.react-calendar__year-view__months__month:not(.react-calendar__tile--hasActive),
+.react-calendar__tile--now.react-calendar__century-view__decades__decade:not(.react-calendar__tile--hasActive), 
+.react-calendar__tile--now.react-calendar__decade-view__years__year:not(.react-calendar__tile--hasActive) {
+  background: none;
+  outline: 1px solid ${transparentize(0.5, colors.blue[500])};
+  outline-offset: -2px;
 
-    .rc-time-picker-panel-select li:focus-visible {
-      outline: 1px solid ${colors.lightBlue[500]};
-    }
+  &:hover,
+  &:focus-visible {
+    background: ${transparentize(0.9, colors.blue[500])};
+  }
+}
 
-    .rc-time-picker-panel-select li:hover,
-    .rc-time-picker-panel-select-option-selected {
-      background: rgba(156,192,162,0.2);
-    }
-
+.react-calendar__tile--now.react-calendar__tile--active {
+  abbr {
+    background: ${colors.blue[500]};
   }
 }
 `;
 
+export const IconContainer = styled(motion.div)`
+  padding: 2px;
+`;
+
+export const CalendarContainer = styled.div`
+  background: ${colors.white[500]};
+`;
+
 export const Container = styled.div`
   display: flex;
-  font-family: 'Lato Regular', sans-serif;
+  column-gap: 5px;
 `;
 
-export const TimeContainer = styled.div`
-  .rc-time-picker-clear {
-    display: none;
-  }
-
-  .rc-time-picker .rc-time-picker-input {
-    ${inputMixin()};
-  }
+export const TimePickerContainer = styled.div`
+  background: ${colors.white[500]};
+  display: flex;
 `;
 
-export const DateContainer = styled.div`
-  font-family: 'Lato Regular', sans-serif;
-  margin-right: 5px;
+export const TimePickerList = styled.div`
+  flex: 1;
+`;
 
-  .PresetDateRangePicker_panel {
-    padding: 0 22px 11px;
-  }
-  .PresetDateRangePicker_button {
-    position: relative;
-    height: 100%;
-    text-align: center;
-    background: 0 0;
-    border: 2px solid #00a699;
-    color: #00a699;
-    padding: 4px 12px;
-    margin-right: 8px;
-    font: inherit;
-    font-weight: 700;
-    line-height: normal;
-    overflow: visible;
-    -moz-box-sizing: border-box;
-    box-sizing: border-box;
-    cursor: pointer;
-  }
-  .PresetDateRangePicker_button:active {
-    outline: 0;
-  }
-  .PresetDateRangePicker_button__selected {
-    color: #fff;
-    background: #00a699;
-  }
-  .SingleDatePickerInput {
-    display: inline-block;
-    background-color: #fff;
-  }
-  .SingleDatePickerInput__withBorder {
-    border-radius: 2px;
-    border: 1px solid #dbdbdb;
-  }
-  .SingleDatePickerInput__rtl {
-    direction: rtl;
-  }
-  .SingleDatePickerInput__disabled {
-    background-color: #f2f2f2;
-  }
-  .SingleDatePickerInput__block {
-    display: block;
-  }
-  .SingleDatePickerInput__showClearDate {
-    padding-right: 30px;
-  }
-  .SingleDatePickerInput_clearDate {
-    background: 0 0;
-    border: 0;
-    color: inherit;
-    font: inherit;
-    line-height: normal;
-    overflow: visible;
-    cursor: pointer;
-    padding: 10px;
-    margin: 0 10px 0 5px;
-    position: absolute;
-    right: 0;
-    top: 50%;
-    transform: translateY(-50%);
-  }
-  .SingleDatePickerInput_clearDate__default:focus,
-  .SingleDatePickerInput_clearDate__default:hover {
-    background: #dbdbdb;
-    border-radius: 50%;
-  }
-  .SingleDatePickerInput_clearDate__small {
-    padding: 6px;
-  }
-  .SingleDatePickerInput_clearDate__hide {
-    visibility: hidden;
-  }
-  .SingleDatePickerInput_clearDate_svg {
-    fill: #82888a;
-    height: 12px;
-    width: 15px;
-    vertical-align: middle;
-  }
-  .SingleDatePickerInput_clearDate_svg__small {
-    height: 9px;
-  }
-  .SingleDatePickerInput_calendarIcon {
-    background: 0 0;
-    border: 0;
-    color: inherit;
-    font: inherit;
-    line-height: normal;
-    overflow: visible;
-    cursor: pointer;
-    display: inline-block;
-    vertical-align: middle;
-    padding: 10px;
-    margin: 0 5px 0 10px;
-  }
-  .SingleDatePickerInput_calendarIcon_svg {
-    fill: #82888a;
-    height: 15px;
-    width: 14px;
-    vertical-align: middle;
-  }
-  .SingleDatePicker {
-    position: relative;
-    display: inline-block;
-  }
-  .SingleDatePicker__block {
-    display: block;
-  }
-  .SingleDatePicker_picker {
-    z-index: 1;
-    background-color: #fff;
-    position: absolute;
-  }
-  .SingleDatePicker_picker__rtl {
-    direction: rtl;
-  }
-  .SingleDatePicker_picker__directionLeft {
-    left: 0;
-  }
-  .SingleDatePicker_picker__directionRight {
-    right: 0;
-  }
-  .SingleDatePicker_picker__portal {
-    background-color: rgba(0, 0, 0, 0.3);
-    position: fixed;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-  }
-  .SingleDatePicker_picker__fullScreenPortal {
-    background-color: #fff;
-  }
-  .SingleDatePicker_closeButton {
-    background: 0 0;
-    border: 0;
-    color: inherit;
-    font: inherit;
-    line-height: normal;
-    overflow: visible;
-    cursor: pointer;
-    position: absolute;
-    top: 0;
-    right: 0;
-    padding: 15px;
-    z-index: 2;
-  }
-  .SingleDatePicker_closeButton:focus,
-  .SingleDatePicker_closeButton:hover {
-    text-decoration: none;
-  }
-  .SingleDatePicker_closeButton_svg {
-    height: 15px;
-    width: 15px;
-    fill: #cacccd;
-  }
-  .DayPickerKeyboardShortcuts_buttonReset {
-    background: 0 0;
-    border: 0;
-    border-radius: 0;
-    color: inherit;
-    font: inherit;
-    line-height: normal;
-    overflow: visible;
-    padding: 0;
-    cursor: pointer;
-    font-size: 14px;
-  }
-  .DayPickerKeyboardShortcuts_buttonReset:active {
-    outline: 0;
-  }
-  .DayPickerKeyboardShortcuts_show {
-    width: 33px;
-    height: 26px;
-    position: absolute;
-    z-index: 2;
-  }
-  .DayPickerKeyboardShortcuts_show::before {
-    content: '';
-    display: block;
-    position: absolute;
-  }
-  .DayPickerKeyboardShortcuts_show__bottomRight {
-    bottom: 0;
-    right: 0;
-  }
-  .DayPickerKeyboardShortcuts_show__bottomRight::before {
-    border-top: 26px solid transparent;
-    border-right: 33px solid #00a699;
-    bottom: 0;
-    right: 0;
-  }
-  .DayPickerKeyboardShortcuts_show__bottomRight:hover::before {
-    border-right: 33px solid #008489;
-  }
-  .DayPickerKeyboardShortcuts_show__topRight {
-    top: 0;
-    right: 0;
-  }
-  .DayPickerKeyboardShortcuts_show__topRight::before {
-    border-bottom: 26px solid transparent;
-    border-right: 33px solid #00a699;
-    top: 0;
-    right: 0;
-  }
-  .DayPickerKeyboardShortcuts_show__topRight:hover::before {
-    border-right: 33px solid #008489;
-  }
-  .DayPickerKeyboardShortcuts_show__topLeft {
-    top: 0;
-    left: 0;
-  }
-  .DayPickerKeyboardShortcuts_show__topLeft::before {
-    border-bottom: 26px solid transparent;
-    border-left: 33px solid #00a699;
-    top: 0;
-    left: 0;
-  }
-  .DayPickerKeyboardShortcuts_show__topLeft:hover::before {
-    border-left: 33px solid #008489;
-  }
-  .DayPickerKeyboardShortcuts_showSpan {
-    color: #fff;
-    position: absolute;
-  }
-  .DayPickerKeyboardShortcuts_showSpan__bottomRight {
-    bottom: 0;
-    right: 5px;
-  }
-  .DayPickerKeyboardShortcuts_showSpan__topRight {
-    top: 1px;
-    right: 5px;
-  }
-  .DayPickerKeyboardShortcuts_showSpan__topLeft {
-    top: 1px;
-    left: 5px;
-  }
-  .DayPickerKeyboardShortcuts_panel {
-    overflow: auto;
-    background: #fff;
-    border: 1px solid #dbdbdb;
-    border-radius: 2px;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    right: 0;
-    left: 0;
-    z-index: 2;
-    padding: 22px;
-    margin: 33px;
-    text-align: left;
-  }
-  .DayPickerKeyboardShortcuts_title {
-    font-size: 16px;
-    font-weight: 700;
-    margin: 0;
-  }
-  .DayPickerKeyboardShortcuts_list {
-    list-style: none;
-    padding: 0;
-    font-size: 14px;
-  }
-  .DayPickerKeyboardShortcuts_close {
-    position: absolute;
-    right: 22px;
-    top: 22px;
-    z-index: 2;
-  }
-  .DayPickerKeyboardShortcuts_close:active {
-    outline: 0;
-  }
-  .DayPickerKeyboardShortcuts_closeSvg {
-    height: 15px;
-    width: 15px;
-    fill: #cacccd;
-  }
-  .DayPickerKeyboardShortcuts_closeSvg:focus,
-  .DayPickerKeyboardShortcuts_closeSvg:hover {
-    fill: #82888a;
-  }
-  .CalendarDay {
-    -moz-box-sizing: border-box;
-    box-sizing: border-box;
-    cursor: pointer;
-    font-size: 14px;
-    text-align: center;
-  }
-  .CalendarDay:active {
-    outline: 0;
-  }
-  .CalendarDay__defaultCursor {
-    cursor: default;
-  }
-  .CalendarDay__default {
-    border: 1px solid #e4e7e7;
-    color: #484848;
-    background: #fff;
-  }
-  .CalendarDay__default:hover {
-    background: #e4e7e7;
-    border: 1px solid #e4e7e7;
-    color: inherit;
-  }
-  .CalendarDay__hovered_offset {
-    background: #f4f5f5;
-    border: 1px double #e4e7e7;
-    color: inherit;
-  }
-  .CalendarDay__outside {
-    border: 0;
-    background: #fff;
-    color: #484848;
-  }
-  .CalendarDay__outside:hover {
-    border: 0;
-  }
-  .CalendarDay__blocked_minimum_nights {
-    background: #fff;
-    border: 1px solid #eceeee;
-    color: #cacccd;
-  }
-  .CalendarDay__blocked_minimum_nights:active,
-  .CalendarDay__blocked_minimum_nights:hover {
-    background: #fff;
-    color: #cacccd;
-  }
-  .CalendarDay__highlighted_calendar {
-    background: #ffe8bc;
-    color: #484848;
-  }
-  .CalendarDay__highlighted_calendar:active,
-  .CalendarDay__highlighted_calendar:hover {
-    background: #ffce71;
-    color: #484848;
-  }
-  .CalendarDay__selected_span {
-    background: #66e2da;
-    border: 1px double #33dacd;
-    color: #fff;
-  }
-  .CalendarDay__selected_span:active,
-  .CalendarDay__selected_span:hover {
-    background: #33dacd;
-    border: 1px double #33dacd;
-    color: #fff;
-  }
-  .CalendarDay__selected,
-  .CalendarDay__selected:active,
-  .CalendarDay__selected:hover {
-    background: #00a699;
-    border: 1px double #00a699;
-    color: #fff;
-  }
-  .CalendarDay__hovered_span,
-  .CalendarDay__hovered_span:hover {
-    background: #b2f1ec;
-    border: 1px double #80e8e0;
-    color: #007a87;
-  }
-  .CalendarDay__hovered_span:active {
-    background: #80e8e0;
-    border: 1px double #80e8e0;
-    color: #007a87;
-  }
-  .CalendarDay__blocked_calendar,
-  .CalendarDay__blocked_calendar:active,
-  .CalendarDay__blocked_calendar:hover {
-    background: #cacccd;
-    border: 1px solid #cacccd;
-    color: #82888a;
-  }
-  .CalendarDay__blocked_out_of_range,
-  .CalendarDay__blocked_out_of_range:active,
-  .CalendarDay__blocked_out_of_range:hover {
-    background: #fff;
-    border: 1px solid #e4e7e7;
-    color: #cacccd;
-  }
-  .CalendarDay__hovered_start_first_possible_end {
-    background: #eceeee;
-    border: 1px double #eceeee;
-  }
-  .CalendarDay__hovered_start_blocked_min_nights {
-    background: #eceeee;
-    border: 1px double #e4e7e7;
-  }
-  .CalendarMonth {
-    background: #fff;
-    text-align: center;
-    vertical-align: top;
-    user-select: none;
-  }
-  .CalendarMonth_table {
-    border-collapse: collapse;
-    border-spacing: 0;
-  }
-  .CalendarMonth_verticalSpacing {
-    border-collapse: separate;
-  }
-  .CalendarMonth_caption {
-    color: #484848;
-    font-size: 18px;
-    text-align: center;
-    padding-top: 22px;
-    padding-bottom: 37px;
-    caption-side: initial;
-  }
-  .CalendarMonth_caption__verticalScrollable {
-    padding-top: 12px;
-    padding-bottom: 7px;
-  }
-  .CalendarMonthGrid {
-    background: #fff;
-    text-align: left;
-    z-index: 0;
-  }
-  .CalendarMonthGrid__animating {
-    z-index: 1;
-  }
-  .CalendarMonthGrid__horizontal {
-    position: absolute;
-    left: 9px;
-  }
-  .CalendarMonthGrid__vertical,
-  .CalendarMonthGrid__vertical_scrollable {
-    margin: 0 auto;
-  }
-  .CalendarMonthGrid_month__horizontal {
-    display: inline-block;
-    vertical-align: top;
-    min-height: 100%;
-  }
-  .CalendarMonthGrid_month__hideForAnimation {
-    position: absolute;
-    z-index: -1;
-    opacity: 0;
-    pointer-events: none;
-  }
-  .CalendarMonthGrid_month__hidden {
-    visibility: hidden;
-  }
-  .DayPickerNavigation {
-    position: relative;
-    z-index: 2;
-  }
-  .DayPickerNavigation__horizontal {
-    height: 0;
-  }
-  .DayPickerNavigation__verticalScrollable_prevNav {
-    z-index: 1;
-  }
-  .DayPickerNavigation__verticalDefault {
-    position: absolute;
-    width: 100%;
-    height: 52px;
-    bottom: 0;
-    left: 0;
-  }
-  .DayPickerNavigation__verticalScrollableDefault {
-    position: relative;
-  }
-  .DayPickerNavigation__bottom {
-    height: auto;
-  }
-  .DayPickerNavigation__bottomDefault {
-    display: flex;
-    justify-content: space-between;
-  }
-  .DayPickerNavigation_button {
-    cursor: pointer;
-    user-select: none;
-    border: 0;
-    padding: 0;
-    margin: 0;
-  }
-  .DayPickerNavigation_button__default {
-    border: 1px solid #e4e7e7;
-    background-color: #fff;
-    color: #757575;
-  }
-  .DayPickerNavigation_button__default:focus,
-  .DayPickerNavigation_button__default:hover {
-    border: 1px solid #c4c4c4;
-  }
-  .DayPickerNavigation_button__default:active {
-    background: #f2f2f2;
-  }
-  .DayPickerNavigation_button__disabled {
-    cursor: default;
-    border: 1px solid #f2f2f2;
-  }
-  .DayPickerNavigation_button__disabled:focus,
-  .DayPickerNavigation_button__disabled:hover {
-    border: 1px solid #f2f2f2;
-  }
-  .DayPickerNavigation_button__disabled:active {
-    background: 0 0;
-  }
-  .DayPickerNavigation_button__horizontalDefault {
-    position: absolute;
-    top: 18px;
-    line-height: 0.78;
-    border-radius: 3px;
-    padding: 6px 9px;
-  }
-  .DayPickerNavigation_bottomButton__horizontalDefault {
-    position: static;
-    margin: -10px 22px 30px;
-  }
-  .DayPickerNavigation_leftButton__horizontalDefault {
-    left: 22px;
-  }
-  .DayPickerNavigation_rightButton__horizontalDefault {
-    right: 22px;
-  }
-  .DayPickerNavigation_button__verticalDefault {
-    padding: 5px;
-    background: #fff;
-    box-shadow: 0 0 5px 2px rgba(0, 0, 0, 0.1);
-    position: relative;
-    display: inline-block;
-    text-align: center;
-    height: 100%;
-    width: 50%;
-  }
-  .DayPickerNavigation_nextButton__verticalDefault {
-    border-left: 0;
-  }
-  .DayPickerNavigation_nextButton__verticalScrollableDefault,
-  .DayPickerNavigation_prevButton__verticalScrollableDefault {
-    width: 100%;
-  }
-  .DayPickerNavigation_svg__horizontal {
-    height: 19px;
-    width: 19px;
-    fill: #82888a;
-    display: block;
-  }
-  .DayPickerNavigation_svg__vertical {
-    height: 42px;
-    width: 42px;
-    fill: #484848;
-  }
-  .DayPickerNavigation_svg__disabled {
-    fill: #f2f2f2;
-  }
-  .DayPicker {
-    background: #fff;
-    position: relative;
-    text-align: left;
-  }
-  .DayPicker__horizontal {
-    background: #fff;
-  }
-  .DayPicker__verticalScrollable {
-    height: 100%;
-  }
-  .DayPicker__hidden {
-    visibility: hidden;
-  }
-  .DayPicker__withBorder {
-    box-shadow: 0 2px 6px rgba(0, 0, 0, 0.05), 0 0 0 1px rgba(0, 0, 0, 0.07);
-    border-radius: 3px;
-  }
-  .DayPicker_portal__horizontal {
-    box-shadow: none;
-    position: absolute;
-    left: 50%;
-    top: 50%;
-  }
-  .DayPicker_portal__vertical {
-    position: initial;
-  }
-  .DayPicker_focusRegion {
-    outline: 0;
-  }
-  .DayPicker_calendarInfo__horizontal,
-  .DayPicker_wrapper__horizontal {
-    display: inline-block;
-    vertical-align: top;
-  }
-  .DayPicker_weekHeaders {
-    position: relative;
-  }
-  .DayPicker_weekHeaders__horizontal {
-    margin-left: 9px;
-  }
-  .DayPicker_weekHeader {
-    color: #757575;
-    position: absolute;
-    top: 62px;
-    z-index: 2;
-    text-align: left;
-  }
-  .DayPicker_weekHeader__vertical {
-    left: 50%;
-  }
-  .DayPicker_weekHeader__verticalScrollable {
-    top: 0;
-    display: table-row;
-    border-bottom: 1px solid #dbdbdb;
-    background: #fff;
-    margin-left: 0;
-    left: 0;
-    width: 100%;
-    text-align: center;
-  }
-  .DayPicker_weekHeader_ul {
-    list-style: none;
-    margin: 1px 0;
-    padding-left: 0;
-    padding-right: 0;
-    font-size: 14px;
-  }
-  .DayPicker_weekHeader_li {
-    display: inline-block;
-    text-align: center;
-  }
-  .DayPicker_transitionContainer {
-    position: relative;
-    overflow: hidden;
-    border-radius: 3px;
-  }
-  .DayPicker_transitionContainer__horizontal {
-    transition: height 0.2s ease-in-out;
-  }
-  .DayPicker_transitionContainer__vertical {
-    width: 100%;
-  }
-  .DayPicker_transitionContainer__verticalScrollable {
-    padding-top: 20px;
-    height: 100%;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    right: 0;
-    left: 0;
-    overflow-y: scroll;
-  }
-  .DateInput {
-    margin: 0;
-    padding: 0;
-    background: #fff;
-    position: relative;
-    display: inline-block;
-    width: 130px;
-    vertical-align: middle;
-  }
-  .DateInput__small {
-    width: 97px;
-  }
-  .DateInput__block {
-    width: 100%;
-  }
-  .DateInput__disabled {
-    background: #f2f2f2;
-    color: #dbdbdb;
-  }
-  .DateInput_input {
-    font-weight: 200;
-    font-size: 19px;
-    line-height: 24px;
-    color: #484848;
-    background-color: #fff;
-    width: 100%;
-    padding: 11px 11px 9px;
-    border: 0;
-    border-top: 0;
-    border-right: 0;
-    border-bottom: 2px solid transparent;
-    border-left: 0;
-    border-radius: 0;
-  }
-  .DateInput_input__small {
-    font-size: 15px;
-    line-height: 18px;
-    letter-spacing: 0.2px;
-    padding: 7px 7px 5px;
-  }
-  .DateInput_input__regular {
-    font-weight: auto;
-  }
-  .DateInput_input__readOnly {
-    user-select: none;
-  }
-  .DateInput_input__focused {
-    outline: 0;
-    background: #fff;
-    border: 0;
-    border-top: 0;
-    border-right: 0;
-    border-bottom: 2px solid #008489;
-    border-left: 0;
-  }
-  .DateInput_input__disabled {
-    background: #f2f2f2;
-    font-style: italic;
-  }
-  .DateInput_screenReaderMessage {
-    border: 0;
-    clip: rect(0, 0, 0, 0);
-    height: 1px;
-    margin: -1px;
-    overflow: hidden;
-    padding: 0;
-    position: absolute;
-    width: 1px;
-  }
-  .DateInput_fang {
-    position: absolute;
-    width: 20px;
-    height: 10px;
-    left: 22px;
-    z-index: 2;
-  }
-  .DateInput_fangShape {
-    fill: #fff;
-  }
-  .DateInput_fangStroke {
-    stroke: #dbdbdb;
-    fill: transparent;
-  }
-  .DateRangePickerInput {
-    background-color: #fff;
-    display: inline-block;
-  }
-  .DateRangePickerInput__disabled {
-    background: #f2f2f2;
-  }
-  .DateRangePickerInput__withBorder {
-    border-radius: 2px;
-    border: 1px solid #dbdbdb;
-  }
-  .DateRangePickerInput__rtl {
-    direction: rtl;
-  }
-  .DateRangePickerInput__block {
-    display: block;
-  }
-  .DateRangePickerInput__showClearDates {
-    padding-right: 30px;
-  }
-  .DateRangePickerInput_arrow {
-    display: inline-block;
-    vertical-align: middle;
-    color: #484848;
-  }
-  .DateRangePickerInput_arrow_svg {
-    vertical-align: middle;
-    fill: #484848;
-    height: 24px;
-    width: 24px;
-  }
-  .DateRangePickerInput_clearDates {
-    background: 0 0;
-    border: 0;
-    color: inherit;
-    font: inherit;
-    line-height: normal;
-    overflow: visible;
-    cursor: pointer;
-    padding: 10px;
-    margin: 0 10px 0 5px;
-    position: absolute;
-    right: 0;
-    top: 50%;
-    transform: translateY(-50%);
-  }
-  .DateRangePickerInput_clearDates__small {
-    padding: 6px;
-  }
-  .DateRangePickerInput_clearDates_default:focus,
-  .DateRangePickerInput_clearDates_default:hover {
-    background: #dbdbdb;
-    border-radius: 50%;
-  }
-  .DateRangePickerInput_clearDates__hide {
-    visibility: hidden;
-  }
-  .DateRangePickerInput_clearDates_svg {
-    fill: #82888a;
-    height: 12px;
-    width: 15px;
-    vertical-align: middle;
-  }
-  .DateRangePickerInput_clearDates_svg__small {
-    height: 9px;
-  }
-  .DateRangePickerInput_calendarIcon {
-    background: 0 0;
-    border: 0;
-    color: inherit;
-    font: inherit;
-    line-height: normal;
-    overflow: visible;
-    cursor: pointer;
-    display: inline-block;
-    vertical-align: middle;
-    padding: 10px;
-    margin: 0 5px 0 10px;
-  }
-  .DateRangePickerInput_calendarIcon_svg {
-    fill: #82888a;
-    height: 15px;
-    width: 14px;
-    vertical-align: middle;
-  }
-  .DateRangePicker {
-    position: relative;
-    display: inline-block;
-  }
-  .DateRangePicker__block {
-    display: block;
-  }
-  .DateRangePicker_picker {
-    z-index: 1;
-    background-color: #fff;
-    position: absolute;
-  }
-  .DateRangePicker_picker__rtl {
-    direction: rtl;
-  }
-  .DateRangePicker_picker__directionLeft {
-    left: 0;
-  }
-  .DateRangePicker_picker__directionRight {
-    right: 0;
-  }
-  .DateRangePicker_picker__portal {
-    background-color: rgba(0, 0, 0, 0.3);
-    position: fixed;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-  }
-  .DateRangePicker_picker__fullScreenPortal {
-    background-color: #fff;
-  }
-  .DateRangePicker_closeButton {
-    background: 0 0;
-    border: 0;
-    color: inherit;
-    font: inherit;
-    line-height: normal;
-    overflow: visible;
-    cursor: pointer;
-    position: absolute;
-    top: 0;
-    right: 0;
-    padding: 15px;
-    z-index: 2;
-  }
-  .DateRangePicker_closeButton:focus,
-  .DateRangePicker_closeButton:hover {
-    text-decoration: none;
-  }
-  .DateRangePicker_closeButton_svg {
-    height: 15px;
-    width: 15px;
-    fill: #cacccd;
-  }
-
-  .SingleDatePickerInput__withBorder {
-    border: none;
-    border-radius: 0;
-  }
-
-  .SingleDatePicker_picker {
-    top: 39px !important;
-  }
-
-  .CalendarDay__default:hover {
-    background: ${transparentize(0.8, colors.green[100])};
-    border: 1px solid ${transparentize(0.8, colors.green[100])};
-  }
-
-  .DayPicker__withBorder {
-    border: solid 1px ${colors.gray[200]};
-    box-shadow: 0 10px 24px 0 rgba(29, 39, 41, 0.15);
-  }
-
-  .DateInput {
-    width: 100%;
-  }
-
-  .DateInput_input {
-    ${inputMixin()};
-  }
-
-  .DateInput_fang {
-    display: none;
-  }
-
-  .CalendarDay__selected,
-  .CalendarDay__selected:active,
-  .CalendarDay__selected:hover {
-    background: ${colors.blue[500]};
-    border: 1px double ${colors.blue[500]};
-  }
+export const PickerInput = styled.div`
+  max-width: 180px;
 `;
