+++ packages/ui-core/src/components/date-picker/index.ts
@@ -1,5 +1,3 @@
 import DatePicker from './date-picker.component';
-import ReactCalendar from './datePicker.component';
 
-export { ReactCalendar };
 export default DatePicker;
