+++ packages/ui-core/src/components/index.ts
@@ -75,7 +75,7 @@ import ScrollWrapper from './scroll-wrapper';
 import Tabs from './tabs';
 
 import AbsoluteTime, { TIME_PICKER_CLASS } from './absolute-time';
-import DatePicker, { ReactCalendar } from './date-picker';
+import DatePicker from './date-picker';
 import DropableContainer, {
   DropableContainerVariant,
 } from './dropable-container';
@@ -167,7 +167,6 @@ export {
   Tabs,
   AbsoluteTime,
   DatePicker,
-  ReactCalendar,
   DropableContainer,
   DropdownListContainer,
   RelativeTime,
