+++ packages/ui-core/src/index.ts
@@ -84,7 +84,6 @@ import {
   AbsoluteTime,
   TIME_PICKER_CLASS,
   DatePicker,
-  ReactCalendar,
   DropableContainer,
   DropableContainerVariant,
   RelativeTime,
@@ -215,7 +214,6 @@ export {
   Tabs,
   AbsoluteTime,
   DatePicker,
-  ReactCalendar,
   DropableContainer,
   RelativeTime,
   TimePeriod,
